+++ tests/parser.scm
@@ -1,14 +1,14 @@
 (test "parser: syntax extension"
       (lambda (t)
-        (add-special! "<>" 'html lips.specials.SPLICE)
+        (set-special! "<>" 'html lips.specials.SPLICE)
         (define-macro (html . args)
           (let ((str (--> (list->array (map symbol->string args)) (join "+"))))
             `(string-append "<" ,str "/>")))
 
         (t.is (eval (read "<>(foo bar)") (current-environment)) "<foo+bar/>")
-        (remove-special! "<>")
+        (unset-special! "<>")
 
-        (add-special! "--" 'dash lips.specials.LITERAL)
+        (set-special! "--" 'dash lips.specials.LITERAL)
         (define-macro (dash x)
           `'(,(car x) . (,cadr x)))
 
@@ -17,5 +17,5 @@
         (t.is (read "(--)") '((dash)))
         (t.is (read "(-- x)") '((dash x)))
         (t.is (read "--x") '(dash x))
-        (remove-special! "--")
+        (unset-special! "--")
         (t.is (read "(--)") '(--))))