+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 1.0.0-beta.4
+### Breaking
+* add/remove-specials and add/remove-repr are now set/unset
 ### Features
 * doc strings on variables
 * fix Environment::has (ref) on variable that are undefined
@@ -9,6 +11,7 @@
 * fix parsing escape slash inside strings
 * don't throw exception on parse error in string->number (return false)
 * fix silent accepting two big code point `#\xFFFFFF`
+* fix custom repr of instances created by define-class
 
 ## 1.0.0-beta.3
 ## Bugfix